feat: drop foam
