+++ package.json
@@ -22,7 +22,6 @@
     "lint": "eslint . --fix"
   },
   "devDependencies": {
-    "@carrotsearch/foamtree": "^3.5.1",
     "@iconify-json/ph": "^1.1.12",
     "@jridgewell/source-map": "^0.3.6",
     "@jridgewell/trace-mapping": "^0.3.25",
@@ -56,7 +55,7 @@
     "rollup": "^4.13.0",
     "rollup-plugin-dts": "^6.1.0",
     "rollup-plugin-swc3": "^0.11.2",
-    "squarified": "^0.1.0",
+    "squarified": "^0.1.1",
     "typescript": "^5.2.2",
     "unplugin-icons": "^0.18.5",
     "vite": "^5.2.8"
