--- src/client/components/treemap-v2/index.ts
@@ -1,1 +1,0 @@
-export { TreemapV2 } from './callee'
